+++ src/main/scala/org/yankov/mso/application/converters/StringConverters.scala
@@ -1,7 +1,7 @@
 package org.yankov.mso.application.converters
 
 import org.yankov.mso.application.model.DataModel._
-import org.yankov.mso.application.model.SearchModel.{Operator, Variable}
+import org.yankov.mso.application.search.SearchModel.{Operator, Variable}
 
 object StringConverters {
   def artistToString(x: Artist): String = x.name
